pages/dwh_02/profesorado/recursos_projecte/requirements.txt
pandas
sqlalchemy
psycopg2-binary
mysql-connector-python
requests
